+++ src/FramedClientConfig.ts
@@ -1,5 +1,6 @@
 interface FramedClientConfig {
     multiClient?: boolean;
+    baseUrl?: string;
 }
 
 export default FramedClientConfig;